events: move geo tests to get_test.js
