+++ test/jasmine/tests/click_test.js
@@ -2,7 +2,6 @@ var Plotly = require('@lib/index');
 var Lib = require('@src/lib');
 var Drawing = require('@src/components/drawing');
 var DBLCLICKDELAY = require('@src/constants/interactions').DBLCLICKDELAY;
-var HOVERMINTIME = require('@src/plots/cartesian/constants').HOVERMINTIME;
 
 var d3 = require('d3');
 var createGraphDiv = require('../assets/create_graph_div');
@@ -1314,197 +1313,6 @@ describe('Test click interactions on a ternary plot:', function() {
 
             expect(evt.clientX).toEqual(pointPos[0], 'event.clientX');
             expect(evt.clientY).toEqual(pointPos[1], 'event.clientY');
-        });
-    });
-});
-
-
-describe('Test click interactions on a geo plot:', function() {
-    var mock = require('@mocks/geo_scattergeo-locations.json');
-
-    var mockCopy, gd;
-
-    var blankPos = [10, 10],
-        pointPos,
-        nearPos;
-
-    beforeAll(function(done) {
-        jasmine.addMatchers(customMatchers);
-
-        gd = createGraphDiv();
-        mockCopy = Lib.extendDeep({}, mock);
-        Plotly.plot(gd, mockCopy.data, mockCopy.layout).then(function() {
-            pointPos = getClientPosition('path.point');
-            nearPos = [pointPos[0] - 30, pointPos[1] - 30];
-            destroyGraphDiv();
-            done();
-        });
-    });
-
-    beforeEach(function() {
-        gd = createGraphDiv();
-        mockCopy = Lib.extendDeep({}, mock);
-    });
-
-    afterEach(destroyGraphDiv);
-
-    describe('click events', function() {
-        var futureData;
-
-        beforeEach(function(done) {
-            Plotly.plot(gd, mockCopy.data, mockCopy.layout).then(done);
-
-            gd.on('plotly_click', function(data) {
-                futureData = data;
-            });
-        });
-
-        it('should not be trigged when not on data points', function() {
-            click(blankPos[0], blankPos[1]);
-            expect(futureData).toBe(undefined);
-        });
-
-        it('should contain the correct fields', function() {
-            click(pointPos[0], pointPos[1]);
-
-            var pt = futureData.points[0],
-                evt = futureData.event;
-
-            expect(Object.keys(pt)).toEqual([
-                'data', 'fullData', 'curveNumber', 'pointNumber', 'lon', 'lat',
-                'location'
-            ]);
-
-            expect(pt.curveNumber).toEqual(0, 'points[0].curveNumber');
-            expect(typeof pt.data).toEqual(typeof {}, 'points[0].data');
-            expect(typeof pt.fullData).toEqual(typeof {}, 'points[0].fullData');
-            expect(pt.lat).toEqual(-101.57, 'points[0].lat');
-            expect(pt.lon).toEqual(57.75, 'points[0].lon');
-            expect(pt.location).toEqual(57.75, 'points[0].location');
-            expect(pt.pointNumber).toEqual(0, 'points[0].pointNumber');
-
-            expect(evt.clientX).toEqual(pointPos[0], 'event.clientX');
-            expect(evt.clientY).toEqual(pointPos[1], 'event.clientY');
-        });
-    });
-
-    describe('modified click events', function() {
-        var clickOpts = {
-                altKey: true,
-                ctrlKey: true,
-                metaKey: true,
-                shiftKey: true
-            },
-            futureData;
-
-        beforeEach(function(done) {
-            Plotly.plot(gd, mockCopy.data, mockCopy.layout).then(done);
-
-            gd.on('plotly_click', function(data) {
-                futureData = data;
-            });
-        });
-
-        it('should not be trigged when not on data points', function() {
-            click(blankPos[0], blankPos[1], clickOpts);
-            expect(futureData).toBe(undefined);
-        });
-
-        it('should contain the correct fields', function() {
-            click(pointPos[0], pointPos[1], clickOpts);
-
-            var pt = futureData.points[0],
-                evt = futureData.event;
-
-            expect(Object.keys(pt)).toEqual([
-                'data', 'fullData', 'curveNumber', 'pointNumber', 'lon', 'lat',
-                'location'
-            ]);
-
-            expect(pt.curveNumber).toEqual(0, 'points[0].curveNumber');
-            expect(typeof pt.data).toEqual(typeof {}, 'points[0].data');
-            expect(typeof pt.fullData).toEqual(typeof {}, 'points[0].fullData');
-            expect(pt.lat).toEqual(-101.57, 'points[0].lat');
-            expect(pt.lon).toEqual(57.75, 'points[0].lon');
-            expect(pt.location).toEqual(57.75, 'points[0].location');
-            expect(pt.pointNumber).toEqual(0, 'points[0].pointNumber');
-
-            expect(evt.clientX).toEqual(pointPos[0], 'event.clientX');
-            expect(evt.clientY).toEqual(pointPos[1], 'event.clientY');
-            Object.getOwnPropertyNames(clickOpts).forEach(function(opt) {
-                expect(evt[opt]).toEqual(clickOpts[opt], 'event.' + opt);
-            });
-        });
-    });
-
-    describe('hover events', function() {
-        var futureData;
-
-        beforeEach(function(done) {
-            Plotly.plot(gd, mockCopy.data, mockCopy.layout).then(done);
-
-            gd.on('plotly_hover', function(data) {
-                futureData = data;
-            });
-        });
-
-        it('should contain the correct fields', function() {
-            mouseEvent('mousemove', blankPos[0], blankPos[1]);
-            mouseEvent('mousemove', pointPos[0], pointPos[1]);
-
-            var pt = futureData.points[0],
-                evt = futureData.event;
-
-            expect(Object.keys(pt)).toEqual([
-                'data', 'fullData', 'curveNumber', 'pointNumber', 'lon', 'lat',
-                'location'
-            ]);
-
-            expect(pt.curveNumber).toEqual(0, 'points[0].curveNumber');
-            expect(typeof pt.data).toEqual(typeof {}, 'points[0].data');
-            expect(typeof pt.fullData).toEqual(typeof {}, 'points[0].fullData');
-            expect(pt.lat).toEqual(-101.57, 'points[0].lat');
-            expect(pt.lon).toEqual(57.75, 'points[0].lon');
-            expect(pt.location).toEqual(57.75, 'points[0].location');
-            expect(pt.pointNumber).toEqual(0, 'points[0].pointNumber');
-
-            expect(evt.clientX).toEqual(pointPos[0], 'event.clientX');
-            expect(evt.clientY).toEqual(pointPos[1], 'event.clientY');
-        });
-    });
-
-    describe('unhover events', function() {
-        var futureData;
-
-        beforeEach(function(done) {
-            Plotly.plot(gd, mockCopy.data, mockCopy.layout).then(done);
-
-            gd.on('plotly_unhover', function(data) {
-                futureData = data;
-            });
-        });
-
-        it('should contain the correct fields', function(done) {
-            move(pointPos[0], pointPos[1], nearPos[0], nearPos[1], HOVERMINTIME + 10).then(function() {
-                var pt = futureData.points[0],
-                    evt = futureData.event;
-
-                expect(Object.keys(pt)).toEqual([
-                    'data', 'fullData', 'curveNumber', 'pointNumber', 'lon', 'lat',
-                    'location'
-                ]);
-
-                expect(pt.curveNumber).toEqual(0, 'points[0].curveNumber');
-                expect(typeof pt.data).toEqual(typeof {}, 'points[0].data');
-                expect(typeof pt.fullData).toEqual(typeof {}, 'points[0].fullData');
-                expect(pt.lat).toEqual(-101.57, 'points[0].lat');
-                expect(pt.lon).toEqual(57.75, 'points[0].lon');
-                expect(pt.location).toEqual(57.75, 'points[0].location');
-                expect(pt.pointNumber).toEqual(0, 'points[0].pointNumber');
-
-                expect(evt.clientX).toEqual(nearPos[0], 'event.clientX');
-                expect(evt.clientY).toEqual(nearPos[1], 'event.clientY');
-            }).then(done);
         });
     });
 });
